+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Don't indent cascades farther than their receiver method chains.
 * Optimize line splitting cascades (#811).
 * Split empty catch blocks with finally clauses (#1029).
+* Split empty catch blocks with catches after them.
 
 # 2.0.1